Try to remove set of locale
There is no en_US.UTF-8 locale in Arch Linux out of the box. Try to remove
explicit set of locale and see what will break. I set it initially, because
some of the servers had russian locale back in the days, but I am not sure this
is still true somewhere.