Remove favicon from cgit
We set root to /usr/share/webapps/cgit for cgit location. This directory
contains favicon which served automatically. One way to stop it is to
remove favicon, but every package update will bring it back in.

So I just hardcoded into nginx config return code of 404 for favicon
request to do not serve it.