Add personal global gitignore file
Usually I have tags file from ctags in root of the repo to simplify code
navigation. Some repos have tags file in gitignore, but some doesn't.
Create personal gitignore files to hide tags file everywhere. This is
not a per repositiry setting, but personal one, so it's fine to
configure it globally.